Updated for the new demo server
Fixed the admin page display to not say that all users have access.
Updated the newRepo script to be relative

Log rotation for the new demo server

Changed the relm to name the repository

Changed the description of the repository to match the actual repository

The admin page was not disabled and we should not say it is.
The password page will redirect to the admin page
if the demo user tries to change the password.

Fixed the domain name to match the demo server

Actually, I want a simple example of a .svn_index file