The demo is now much more than just that.  With independant repository access control behavior and individual admins, this can be used now for a wide range of needs.
There are now 4 sets of browser tests that are there mainly to help
show the specific cases of more complex client-side technology that
may need to be looked at as far as supporting behaviors for Insurrection.

The authentication files themselves are now hiden elsewhere (different
repository) so as to not expose the details and list of users publicly.