Removed the kludge mouse-over trick to fix up the URLs by having the mod_rewrite trick on the server do it for me.
This also lets the SVN Index pages advertise their RSS feeds as the
URL can be pre-determined for them.  (It is just the repository URL
with the appended "?Insurrection=rss" query string.  How nice is that?)

Note that this does have at least one problem in that any path with
a "?" in it will fail to work correctly as the mod_rewrite code does
not deal with it correctly.

See http://issues.apache.org/bugzilla/show_bug.cgi?id=34602