Change the index.html into a dynamic page that reads the repository descriptions and builds the page about the repositories.
Change .htaccess to make index.cgi the index page.

Added the auth_index.cgi that requires authentication such that it
can then show the list of repositories that are available to the
specific user.

index.html became index.template and it is now the template that is loaded
into the page. Note that the <repos/> tag in the template is where the
dynamic repository table is to be inserted.

Initial work on getting user-creator tracking - that is, tracking of which
user created which other user. This is useful for limiting the users
seen on the admin page to those users the admin directly had created.
This is only very preliminary work but it would be best to actually support
the updated password file format first.