A refactoring of the code to remove common.pl as it was only used from admin.pl and thus should actually be part of admin.pl, saving an extra file load stange.
Headers for web pages now have expire times that are more reasonable for
some of the pages so as to not force a fresh page load every time. The
default in the admin.pl (common code) is now 1 day with a few pages
providing different values.

The code structuring also now uses (for the most part) the CGI module's
header method to build the headers. I normally have not used the CGI
module for output but for this case I am doing so.

Fixed up a number of places where the server was called Subversion when
it really should have been Insurrection.