Ok, so, for those 1% of non-Mozilla/non-IE based browsers, I just spent some time making Insurrection "mostly work" for them.
Here is what has been done so far:

1) The revision history (log.cgi) detects known broken browsers and
causes the output to become plain-jane HTML. This is handy for
users of Opera and Konqueror as they were completely broken.
Safari is also listed because it has some problems with XSLT based
pages and all Safari versions prior to OS-X 10.3.9 did not handle
any XSLT.

2) The SVN directory browsing is detected and handled via a web page
that does all of the access requests via XMLHttp object and the
generation of the HTML DOM via JavaScript. Most of this was already
in place due to the support for the in-line directory expansion.
Thus this change was relatively easy to implement with no overhead
to those XSLT supporting browsers. However, this does depend on
the XMLHttp object support in the browser and there are some older
browsers that do not support this. This requires that browsers
pass the Browser Test #4, which can be accessed at:

http://svn.sinz.com:8000/browser-tests/test4.html

or can be found in the Insurrection source code.
Most of this trick is done by having the logic within the
mod_rewrite directives and then just a very simple CGI which
outputs the correct HTML for the page. (It is almost completely static)

Supporting browsers that do not pass test #4 or test #1 will
currently not be attempted due to the adverse impact it would
have on the browsers (99%) that do support this.

In addition, rather that assuming that browsers fail and listing the
known working ones, I am currently assuming that browsers work and
only listing the known failing browsers. This is because of the
major advantages of using XML/XSLT on the client side.


Note that some minor changes were made to clean up some of the
generated results just to keep everyone on their best behavior.