Better support of the special character test cases in the JavaScript code that loads the XML from Subversion and builds the display.
Note that some browsers do not handle all of the test cases correctly to
the server, but that is outside of the scope of this problem.

Also made a minor change such that a directory with the name "_" can not
confuse things.  It was silly to use "_" as a special character when we know
that "/" is, by definition, a special character in a file path name.