Some major restructuring of the way Insurrection handles the CGIs. This was done mainly to reduce the overhead of the authentication checks but it has the nice side effect of making all repository URLs consistent with the Subversion repository URLs with some CGI query parameters added.
For example, if /svn/Web/home/ is the home directory in the repository,
then /svn/Web/home/?Insurrection=log is the revision log of that
directory.

This is all done with some trickery within the Apache mod_ReWrite engine
and the Apache mod_proxy.

The security changes are due to the fact that mod_authz_svn verifies the
path access rights *before* mod_ReWrite (and thus before the proxy rule)
gets to modify it.  Thus, whatever the access rights or lack of rights
are, the are already enforced before the proxy event and thus the target
script has no need to check for authentication.  The CGIs now will try
to prevent direct execution by redirecting to the Subversion repository
URLs if they notice that they are not being accessed via the proxy process.

This allows the reduction in overhead for each script (no need to check
the authentication rights and no chance of getting it "wrong" relative
to what mod_authz_svn does).  It also lets us remove the auth_*.cgi
stubs that were used to provide forced login versions of the scripts.

The Log History page has been fixed to correctly request the next
history page by following the history of the item.  This was actually
working and then broken 10 days ago when I first tried to get the
URL paths cleaned up and did not remember exactly why it was so hard to do ;-)

All images (HTML <img> tags) now have the alt attribute set, even if it is
just the empty string.  This is considered required for HTML 4.01
validation and I had a few that I forgot to include.

All links to other scripts are now handled via the ?Insurrection=XXX
rewrite/proxy rule so there were a number of changes to the .htaccess
file.  (And lots of comments)  This also removed a lot of configuration
information from the Insurrection.js file.  This is a good thing.
However, the number of configuration points within .htaccess has
increased because of this.

The RSS feed now is a bit cleaner - one of the links was not correctly
escaped.  The target of the RSS entry links is the LOG CGI.  The LOG
cgi has been updated to only use the svnlook overhead when needed.
Since the RSS feed links to a specific revision, this makes the
extra server load for RSS clickthrough much, much lower, especially on
repositories that are large (as in many revisions).

The Annotate/Blame feature now lets you click on any specific revision
and go directly to that log entry.  I am still not sure if that is the
best design but for now, we will stick with that.  Maybe having a
popup of the log entry would be better.

The Annotate/Blame CGI now also produces better HTML - it is fully
HTML 4.01 verified.