Fixed up some HTML errors in the history display (I had a <div> inside of a <span> - silly me :-)
Added some XSLT logic to split the user name at the "@" and to use an image
for the "@" to help reduce the ability of scrapers.  Not that it would help
much as the XML would still have the real user name and a smart scraper could
still handle the image case, but might as well try.  The real reason I did
this is bacause I wanted to have the user name split on two lines when it
includes EMail addresses.