Oh boy - what a HACK!  This is one for the ages!
There is this problem with the Apache mod_rewrite engine and certain
characters within the URI.  It basically messes up (either in the CGI
parameters if you use "redirect" or in the URI if you use "proxy")

Anyway, the problem is bad enought that there is no perfect workaround
but there is a hack.

We take what was going to be part of PATH_INFO and put it, insted, as
the *last* CGI parameter to the proxy (yes, proxy is important here)
request.  But this would not be enough as certain characters will
have already been un-escaped since the path part of the URI is 
un-escaped before going into rewrite, thus causing all sorts of problems.

So, on the server side, the admin.pl (which is the first thing included
by all pages and which does the actual setup the the CGI perl module)
does some rebuilding and reparsing of the URI in order to undo the whole
mess.  It even puts the PATH_INFO back where it should be (it actually
has to since we have an already de-escaped path which can not be processed
by the CGI module's query string handling - again, due to the mod_rewrite
problem)

So, with the combination of the tricky perl code and the strange mod_rewrite
proxy operation and the interesting way Apache and CGI work, this actually
gets all but one of the test cases working.

The one test case that does not work is that of a double-space as the
first space within the PATH_INFO of the URI.  (That is, it was %20%20
but due to the escaping problem in mod_rewrite it becomes 2 spaces which
due to the HTTP parsing becomes only 1 space.)

This is only an issue if it is the first space in the URI.  If there is
another single space before the double space then this is not an issue.
(The parsing rules are somewhat simple)

Anyway, this is the best I can do without fixes from Apache.  And when
the fix comes from Apache, we would want to undo all of this.