The last release was at revision 111, so this is 22 revisions later...
The major visible items in this release are:

1)  URLs now always look like Subversion repository URL paths with just
    some extra CGI parameters added to them.  This is all possible since
    the GET method is not used for WebDAV protocol.  This also makes the
    interface a bit cleaner for the user.

2)  The work-around to the Apache mod_rewrite bug with certain characters
    and excaping/lack of excaping.  This became even more critical due to
    item #1 above but it was already somewhat critical before.

Minor items:

Better RSS feeds and better support of RSS feeds of sub-sections of the
repository (all the way down to the individual file level if so wished).

User names that are EMail addresses are now displayed a bit differently,
mainly for visual clarity but also to help reduce the scraping of addresses
by spammers.  (Hey, it is a very minor win since the XML already is
trivial to parse and this only is an issue for "open" repositories)

We also got the official code-host.net service up and running (not yet open
for commercial business but well on the way) so there were some changes for
that.