This is one major change!
I moved the page framing stuff to its own directory, just for easier management.

The blank.gif is not 1x1 and not 15x15. There is a spacer.gif that is the 15x15
spacer that is needed for the directory listings.

I added images and a style sheet for doing a "fuel gauge" type of display.
This is used in the first cut of the bandwidth usage page.

There is a new bandwidth usage page that can show to administrators the
breakdown of bandwidth usage for the past month.

There is a new dump CGI that can download dumps of the repository with
and without binary deltas (1.1 vs 1.0 format)

A number of cleanups in the style sheets and other code, mainly to make
things a bit cleaner. As I said, this was a major commit from a
development branch of code.