Marked all regexp that can be marked as "static compiled" as such. For those in loops, this may be a noticable win.  For those not in loops, it will be a win if I ever spend the time to move to mod_perl...
Also made use of "our" to declare variables - this requires Perl 5.6 or
better.  

Removed the use of "ls" from the bandwidth.cgi in admin mode.  I was
just being lazy and, well, today I was not as lazy :-)

I wonder if this project has gotten big enough to warrent making a Perl
module?  I am thinking that is may be there (it started as a small little
thing and now...)  Anyway, time to start thinking about that...