Ok, so someone (who shall remain nameless) has poked me about Atom support. I don't fully understand why we can't all just get along, but, since we can't, I figured that I may as well support both worlds.
Now, this had a bigger impact that you may have guessed as all places where
an RSS feed is available we now needed an Atom feed too.  New icons, new
code to generated the Atom feed, etc.

A lot of the code was grafted from the RSS feed code.  I have not (yet)
refactored the common stuff since there are just enought subtle differences
between Atom and RSS that I was more interested in just "getting it working"

So, now, Insurrection also has Atom feeds in addition to RSS feeds.
They contain almost the same information (within the scope of RSS vs Atom)
and both validate as valid feeds (see links on main page)

Note that I have not built an XSLT for the Atom feed such that it is
viewable in the browser.  It is getting late and I want to get some
sleep, so that will be for another day (if at all)

Also note that I do not use Atom myself so the testing of this is still
relatively raw/minimal.  The feed validates so it should be relatively
safe everywhere.