This is a major change to the look and feel of the administrative pages. All part of the effort to get the rest of the administrative functions completed.
Repository administration is now cleanly managed. Users can be
added at any time and they will only be for the given repository.
Deleting users does not actually delete the account but it does
delete the user from the repository management list.

Lots of image/background/style tricks to make things look nice.

I removed the large background image from the pages as it was
both distracting and unneeded.  This should make the pages a
bit "cleaner" looking all by itself.

Cleaned up the ugly buttons that were being used for login and
password changing.  They were also non-obvious buttons.  I have
fixed that now.

The old management code is still there but it needs to be cleaned up
to be just system administrator features. (This will be where new
repositories are created, etc.)

Note that the hope is that at some point we can do sub-directory
administration, thus the use of the whole tag, not just the repository.
This is not critical right now.

Also, there is a wish to be able to enable/disable the tag hooks.
This will be added after the create repository process is complete.

So, I hope everyone enjoys the new look - I think it is clean and
refreshing.