Some image cleanup - the light green in the table headers was just a bit too light (we can always get it back if we want - the beauty of Subversion :-)
I also added the images and stylesheet for the bold frame version of the table.
This is not being used at this time but I just wanted it symetry.

Added equal padding to the left title when the login or password button is
on the right side.  In addition, system administrators now have the system
administration button on the left side of the title of the first repository
listing table (the same line where the password button is on the right side)

Every image was tweeked a bit - just to clean things up.

After this, it is getting time to clean up the main css - removing some
of the old layout items that we no longer use (well, currently the manage.cgi
which was the old admin.cgi still uses some but that will change and then
we can finally clean up the mess.)