For repositories where the system administrator has no access rights, just list them but do not provide RSS/ATOM links nor links to the repository itself since the system administrator has no access rights.
Note that this is just to keep things clean as we don't want to give
blanket rights to the administrator.  However, the administrator does
have the right to change access rights which would then show up in the
administration logs and thus there would be a full audit of if the
administrator looked at any files within a private repository.

(In the above, Administrator is the administrator of the server, not
of a specific repository or group of repositories)