Added support for editing the repository description field in the repository administration page.  I am not sure I really want this to be available to general administrators but for now I will make it generally available. This uses a cute trick to help filter bad HTML by pushing the text into a <span> innerHTML and then reading it back out when saving.  This way any malformed HTML will be "fixed" (either stripped or closed).  A nasty hack but it does work...
Removed some of the bandwidth details pages - just the kept the display
in the repository admin page and then the raw details pages.

Some minor cleanup of styles that are not in use.  Some are "still" in
use but in the "dead" manage.cgi which will be replaced really soon now.