This is a relatively large commit due to the amount of things touched.
First, I cleaned up some of the way tables and frames are managed such
that I can produce some visuals that I wanted to do.

I added a client-side tabbed-page feature that works rather well.  This
comes from some of my bag-o-tricks.  I have noticed some rendering problems
in Opera (what are those white lines?) but all other browsers do render it
correctly.  I also noticed a layout quirk in Safari, but it is only a quirk
and may only be noticeable to me :-)

The tabbed page display now let me make a nicer main index page.  This is
just the first cut at that, but it does work.  Since the index.template
is a customization point, I will not worry too much about it.

Lots of CSS changes to the innerframe/frame.css file to add in the tabbed
features.  I am sure this could have been done in a nicer manner but this
works in all browsers (too bad IE has so many CSS limitations - that is
where most of the problems are...)

The bandwidth display code for admins now does even more regex tricks on
the page in order to take advantage of the tabbed-page tricks.  I think
it looked so much better.

Ok, when I talk about "better looking" that is within the context of my
artistic abilities with GIMP.  That is to say, it may not actually look
better to others.  But I have tried my best to not make it look worse. :-)