Put in checks to prevent direct access to the *.template files
Updated the innerframe style sheet to make paragraphs have a reasonable
font size within the frame.

Updated the index.template to depend on the new style sheet.

Don't show the detail bandwidth usage button if there are no details
to see.

General cleanup in the bandwidth CGI - removed stuff that was no longer
used (should have been removed before) and added a check for the
no-stats.flag in a repository.

The tabs.js now supports a URL targetting a specific tab.  This lets
a page make a link to a specific tab number.  (Yes, I did it by number)
A correct target will override any cookie that may be set.