Removed the display of the $Id$ string in the footer of the pages. The $Id$ string is now in a title attribute of the footer and will show up when the mouse is over the footer.  No need to clutter the page with the $Id$ element.
Added a DOCTYPE and DTD link to the log/history XML.  Also created
a DTD to match the log.  Note that this is very simular but not
the same as the SVN LOG DTD.

This means that the svn and log XMLs are not fully validated XML, so
lets put the valid XML 1.0 icon on those pages.

Note - I did make an RSS DTD but am not going to include it as RSS
does not really fit into a DTD due to the lack of required order.
This is actually a good thing (IMHO) since order should not matter.
Plus, the general view seems to be that RSS does not have a formal
DTD.  So, no DOCTYPE for RSS feeds.

The same seems to be true of Atom feeds.  Until the IETF makes a formal
DTD for Atom feeds, I will not have a DOCTYPE for Atom feeds either.