For a while now I have been thinking about the best way to handle the creation of new repositories in the system.  I have this shell script that does it and I have this web interface too.  I never released them as they were somewhat specific to the way I do business...
Ok, so I am committing this publicly even though it is not template driven
yet...  Maybe it does not need to be.  Anyway, I have now released the
web interface that only Insurrection System Admins can access such that
they can create new repositories in the system.  As far as I can tell,
this will work on any "unix-like" system since it assumes path
characteristics that do not exist on Windows.  (Yes, Windows is the lone
"modern" platform where paths and directories are so very different)

I still want to template this but it gets used so rarely that, well,
the return on the investment of time is just not there.  (Now if SVN
could only "svn copy" from one repo into another...  Templates would
be "trivial" then :-)