A nice improvement to the system from Paul Koning - use head to trim down the list of revisions from svnlook when getting a log history.
This reduces the perl overhead for the operation but does not actually
reduce the svnlook overhead.  If only svnlook was able to support a
revision count (since just number math does not work except in whole-tree
conditions)

Anyway, this can have a significant improvement for large (many, many thousands
of revisions) repositories.