This is part 1 of a 2 part change that will optionally enable the use of enhanced directory listing output.
This part implements the XSLT and JS code needed to deal with the optional
extra information.  It also includes the quicky style sheet entry for it.

I am releasing this due to requests on line about the feature.  This
change here only provides the basic functionality.  The next commit
will provide the actual code to generate the enhanced XML schema.