Ok, so here is the magic that makes the enhanced directory listing information work.  The trick is that a rewrite rule is used to catch the "GET" operation for a directory and rather than let the mod_dav_svn answer it, a new "ls.cgi" is provided that answers the call.
Note that this new CGI is higher overhead than doing this within
the mod_dav_svn.  In fact, if mod_dav_svn were ever modified to
to include this extra information in its output, this rewrite rule
and ls.cgi would not be needed and the rest of the display code
would just continue working as is.

If you don't want to use the enhanced output, just disable the new
rewrite rule (it is commented in the .htaccess file) - you can
still get the enhanced display manually on a per-directory basis
by adding the CGI query ?Insurrection=ls to the end of the path.
(Not that one would normally want to do this manually)

The XML Schema is very simular to the one that mod_dav_svn generates today.
In fact, it is done like this such that an updated mod_dav_svn would
be possible in a backwards compatible way and yet provide the extra
information that so many people want to see in the directory listing.

If the mod_dav_svn would ever grow this extra information output then
this commit would no longer be needed.