There was a problem with dealing with deleted/replaced files and the get operation.  The "-r revnum" option works against the head revision to find the file but if the file has been replaced or deleted, it no longer exists in the head.  To get a specific point to start the search, a peg revision syntax is needed.
This happens to also be true for the blame operation.  Only for blame one
needs to use both the "-r revnum" and the peg revision to get the
correct blame behavior.

Thanks to Paul Koning for noticing the problem with get.cgi and the fix.
It was what I needed to address the blame problem that I had been working
on.