If there is no access file (no access rights defined) then the Insurrection code will assume blanket read-only access.
I am not yet sure if this is a good thing, but it only happens
if the access file does not exist.  If the access file exists
but is empty then no blanket rights are given.

Note that this effectively provides browsing of the repositories
on the server as the code looks for all of the repository directories
within the $SVN_BASE and adds them to the list of available repositories
and thus they get filled into the index.template as if there
was an access file that described the repositories and the access
rights.