If, for some reason, the machine hosting Subversion/Insurrection has user access and the user wishes to do something bad (work around the access security of the repository) there was a chance that the correct building of special headers could get some information from the repository by subverting the security checks on the CGI code when accessed from the local machine.  The CGI code does these checks to make sure that the locally looped proxy event came through the formal URL which would have been access control checked by the mod_authz_svn module.
Remote users could never fake this since it requires going though a local
proxy on the server (unless a proxy was set up on the local server)
but a local user could fake the proxy headers.  Now, it is still possible
to fake the headers but not as easily if the server is doing named
virtual hosting since the host name must match just to get to the code.

In other words, the minor fix here may not actually have fixed anything
but then again, it may not be any more broken than would ever be the
case if a user has an account on the machine and the skills build requests
that look like valid proxied requests.  Better safe than sorry...

(Thanks again to Pete Krawczyk for recommending the \Q...\E construct)