Major new feature:  on-line/web-based editing for the svn:log revprop!
This new feature allows for the on-line editing/updating of the log
messages within the repository.  This only gets enabled for repositories
that have the revprop update hook enabled (since otherwise revprops are
locked)

Currently the code prevents a user from updating a log message unless
the user is logged in as the user who initialy committed the revision
or the user is an administrator for the repository.

This uses some XMLHTTP code and "DHTML" in order to make this work
reasonably seemlessly within the respository log display.

I think I have handled all cases and issues within the code.  The
security was the tricky bit since it also requires write access
to the repository for the user in order to allow for the save of
the new log message.

Again, this only will show up in repositories that have svn:log editing
enabled for the repository.