The log.js XMLHTTP code was not setting the content-type for the request and thus did not work in all configurations.
The savelog.cgi code did not prevent a security hole in that passing
a log message that looked like an option to "svn propset" would not
be prevented.  The only real option that did something was "-F/path/name"
but that is enough.  Anyway, the "--" end-of-options option is now
used before the log message and URL get passed.  Much, much safer.

If you implement r280 you will want to make sure you also get r282
for security reasons (and r281 to keep the dtd up-to-date)