Removed the .htaccess trick for dealing with broken XSLT browsers. This is now handled all in the perl CGIs since even the directory listing is now down in the perl code due to the wish for more details than the mod_dav_svn XML provides.
The main reason to do this is to be able to more completely support
browsers that used to have broken XSLT but are now fixed.  (Or soon
will be)

This specifically is about the Apple Safari browser/webkit.  The
current codebase (420 ) of webkit passes all of the XSLT tests that
I have thrown at it.

This commit just enables the future enahancements to the server-side
browser useragent string checking for broken XSLT support.  Since the
Safari/webkit code that has the fixed XSLT support is not yet
released, I am waiting for final testing before releasing that part
of this update.