The latest beta Safari/AppleWebKit deals with XML/XSLT correctly now so I have enabled that browser as of WebKit 420+  This final version number may change yet but I was assured that it should be safe.
Changed the way the JavaScript tells the server that we really want the
XML format even if the browser is broken.  Rather than adding an extra
CGI parameter to the URL we are now using a HTTP header.  This lets the
default caching behaviors work a bit better and it is really cleaner.

I was going to try to set the User-Agent header but not all browsers
support this.  Specifically, the "broken" Safari browsers did not allow
the setting of this header and yet that is exactly where it was needed.
(Non-broken browsers did not need it but happened to work - go figure!)

Anyway, now the header "X-Insurrection-JS: true" means that the request
came from the JavaScript and to not filter based on the list of known
broken browsers.