Just a minor change to have the repository revision displayed when a specific revision is asked for via the "undocumented" URL syntax to the repository browsing URLs.
Note that this is "undocumented" because it is not a complete
solution.  There is a lot of "other" work needed in order to
completely address the ability of browsing the repository at
any revision.  But for those that may wish to look into this,
it is at least a reasonable requirement to show the revision
when not actually "head"