Browsing a repository at a specific revision was only partially implemented. You could manually enter a URL and see that one page at a specific revision.
I have now updated the XSLT and the JavaScript to support the browsing
feature more completely.  You still need to manually enter a URL to get
into this mode but once there it will correctly pass along that information.

I still don't know how best to hook this up into the core UI.  Somewhere
one may wish to select to browse the repository at a specific revision.
My guess is that would be in the history log display somewhere, but I
don't know where yet.

To try this out, go to:

	http://svn.code-host.net/svn/Insurrection/?Insurrection=ls&r=200

This will browse the public Insurrection repository as it existed at
revision 200.