Well, I now have a place where you can click to start browsing the repository at a given revision.  Since all of the browsing code was working now I figured I may as well find a place to hook it up.
The current location is in the revision log display.  Clicking on the
directory folder (which should show the help text of "Browse repository
at revision xxxx" will switch back to the repository browser based on
the given revision.

There is a minor problem due to following path histories but except for
that this should work just fine.  (The path history thing is a problem
due to the fact that all of this UI change only happens in the XSLT
and there is no understanding of path histories - but for most cases
this will work just fine - as in 95+% of the time)