In order to be a bit safer, we now have all of the rewrite rules only run against the GET and POST operations, just in case one would hit for a DAV operation.
This is actually works around another problem that is due to the way
in which the JavaSVN client sends up the HTTP requests - putting the
port number onto the end of the host name even for the default port.
This caused problems due to the way I do the extra security check
against the host name.

This does not address the problem of that same client talking via HTTPS
and forcing the :443 port - but I still don't know the best generic
solution for that one yet.