Fixed the amazingly stupid mistake in the checking of the user name for making new users.  Wow!  The check basically so bad that it was hard to find a name that was not illegal!
The conditional should have been "&&" to check both the characters *and*
the length limit.  I had it as "||" which basically made it such that
it would only prevent names that were both too long and invalid rather than
either too long or invalid!

(What a blunder...  where is that paper bag...)