Ahh, I figured out what the problem was with rendering of the tests in newer (Firefox 1.5+ and Seamonkey 1.0+) browsers.
It turns out that there is a minor bug in the browsers that causes them
to treat CSS tag selectors as case sensitive if the HTML was generated
via XSLT.  This is new behavior as of Firefox 1.5 vs 1.0.x.

I have fixed the core CSS to work around this bug and have added a test #6
that shows this bug's behavior.

See http://bugzilla.mozilla.org/show_bug.cgi?id=327902 for more details
as to the bug.