When doing annotate of a file, the links to show the log entry for any given revision of the file were too specific.
Due to a lack of pegrev support in svn log, if the revision was at a point
where the file did not exist at the same name, the link would fail.

Since looking up log entries for the repository does not need full path
information, we just look for the log entry within the repository root.
That way it will always work.