The RSS Feed Validator now complains about the missing GUID element. Note that the GUID element is not required by the specification but the Feed Validator claims that some RSS readers may have problems without it. GUID turns out to be just like LINK in this case, so it was an easy fix.  This gets rid the the warning.  The feed format was and still is fully valid as per the RSS 2.0.1 specification.
Note that the Feed Validator still warns about the HTML in the
description of the repository.  This is because some entities are
being escaped and because the Insurrection repository description
contains a link in it.  The RSS generator does not, in and of itself,
cause this problem.  However, it is also not illegal to have
markup within the description.  It is just that some RSS readers may
not handle such markup and as such the validator warns about it.