Some restructuring of the dynamic DOM manipulation in the svnindex.js
One change was to enable the display of an animated image that shows
that the folder is being opened/the system is busy.  This is on a per-folder
basis.

The fold/unfold image/button is now only available to the user when the
folder is not busy.  This removes the confusion as to how the folder
loading works.

The DOM manipulation code now builds all of the elements for the DOM and then
links it into the DOM after it is complete.  This has a significant impact on
the dynamic layout engines of various browsers since DOM elements are not
processed before they are fully complete.

Together these changes make for a slightly cleaner behavior during dynamic
loading.  (Not that the earlier code did not work correctly, but it did
push certain browsers to use more CPU than they really needed to and, in
the world of HTML/DOM/JS you really are limited by the browser capabilities
and the "lowest common behavior")