Nicolas Goy Kuon pointed out that there was no charset defined in the HTTP headers for HTML mime types.  I had over looked this partially due to missing this and partially due to now knowing the best charset to put there.
The problem is that some pages include content that users may have entered
and I can not assume any specific charset other than the charset the user
uses.  So, in stead of making a choice, I just fully "forgot" to make one.

Anyway, this is incorrect and needed to be set, even if wrong, as otherwise
the browser would "randomly pick" (well, random is a bit strong since it is
not random at all for a given configuration but that configuration is not
controlled by the server.)

Anyway, this may make some differences to output such a DIFF and BLAME since
that content includes user provided content from the repository.  However,
in general, this is a much better situation than undefined behavior.