For the ATOM and RSS feeds, the "title" of each feed item was just the revision number (well, actually, "Revision nnn")
However, for many feed systems, this produces useless feed headers as there
is no summary of the commit.

What I have done is to take the first line from the commit message and
included it within the title of each item.  This make the title usually
much more useful than just a revision number.